refactor dialogues to use dialogue template
